Check if count arg is not overflown in MPI_Isend
Call to MPI_Isend use int type to specify the quantity of data to
transfer. As it is usually 32bits, it can easily lead to integer
overflow. This patch aims at checking that the MPI_Isend will not
generate integer overflow, and if it will it shows an informative error
message to the user and stop the program.

Further developement might try to fix this problem, but it is consired
low priority at the moment.